Fix constructor prototypes.

+++ src/FadeLED_Exp.cpp
@@ -10,8 +10,8 @@
 // function.
 FadeLED_Exp::FadeLED_Exp(
     const byte pin, 
-    const unsigned int onTime, 
-    const unsigned int offTime,
+    const unsigned long onTime, 
+    const unsigned long offTime,
     const bool invert
 ) : FadeLED(pin, invert), m_level(0.0), m_output(0)
 {
