+++ src/FadeLED_Lin.h
@@ -34,8 +34,8 @@ class FadeLED_Lin : public FadeLED
          */
         FadeLED_Lin(
             const byte pin, 
-            const unsigned int onTime, 
-            const unsigned int offTime,
+            const unsigned long onTime, 
+            const unsigned long offTime,
             const bool invert = false
         );
